requirements.txt
numpy==1.21.5
nibabel==3.2.1
SimpleITK==2.3.1
Pillow==9.0.1